style: Onboarding button colors and background color change
