+++ lib/View/screens/authentication_ui.dart
@@ -1,6 +1,6 @@
 import 'package:authentication_ui/common/common.dart';
 import 'package:authentication_ui/router/router.dart';
-import 'package:authentication_ui/screens/fade_animationtest.dart';
+import 'package:authentication_ui/View/screens/fade_animationtest.dart';
 import 'package:authentication_ui/widgets/custom_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
@@ -92,7 +92,7 @@ class _AuthenticationUIState extends State<AuthenticationUI> {
                             ),
                             backgroundColor:
                                 const WidgetStatePropertyAll(Colors.white)),
-                        child: const Text(
+                          child: const Text(
                           "Register",
                           style: TextStyle(
                               fontSize: 15,
