+++ lib/View/screens/new_password.dart
@@ -1,46 +1,21 @@
-// ignore_for_file: avoid_print
-
 import 'package:authentication_ui/common/common.dart';
 import 'package:authentication_ui/router/router.dart';
-import 'package:authentication_ui/screens/fade_animationtest.dart';
+import 'package:authentication_ui/View/screens/fade_animationtest.dart';
 import 'package:authentication_ui/widgets/custom_widget.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:pinput/pinput.dart';
 
-class OtpVerificationPage extends StatefulWidget {
-  const OtpVerificationPage({super.key});
+class NewPasswordPage extends StatefulWidget {
+  const NewPasswordPage({super.key});
 
   @override
-  State<OtpVerificationPage> createState() => _OtpVerificationPageState();
+  State<NewPasswordPage> createState() => _NewPasswordPageState();
 }
 
-class _OtpVerificationPageState extends State<OtpVerificationPage> {
+class _NewPasswordPageState extends State<NewPasswordPage> {
   @override
   Widget build(BuildContext context) {
-    final defaultPinTheme = PinTheme(
-      width: 56,
-      height: 56,
-      textStyle: const TextStyle(
-          fontSize: 20, color: Colors.grey, fontWeight: FontWeight.w600),
-      decoration: BoxDecoration(
-        border: Border.all(color: Colors.grey.shade800),
-        borderRadius: BorderRadius.circular(20),
-      ),
-    );
-
-    final focusedPinTheme = defaultPinTheme.copyDecorationWith(
-      border: Border.all(color: Colors.grey),
-      borderRadius: BorderRadius.circular(8),
-    );
-
-    final submittedPinTheme = defaultPinTheme.copyWith(
-      decoration: defaultPinTheme.decoration?.copyWith(
-        color: const Color.fromRGBO(234, 239, 243, 1),
-      ),
-    );
-
     return Scaffold(
       // backgroundColor: const Color(0xFFE8ECF4),
       body: SafeArea(
@@ -68,14 +43,14 @@ class _OtpVerificationPageState extends State<OtpVerificationPage> {
                     FadeInAnimation(
                       delay: 1.3,
                       child: Text(
-                        "OTP Verification",
+                        "Create new password",
                         style: Common().titelTheme,
                       ),
                     ),
                     FadeInAnimation(
                       delay: 1.6,
                       child: Text(
-                        "Enter the verification code we just sent on your email address.",
+                        "Your new password must be unique from those previously used.",
                         style: Common().mediumThemeblack,
                       ),
                     )
@@ -89,31 +64,31 @@ class _OtpVerificationPageState extends State<OtpVerificationPage> {
                     children: [
                       FadeInAnimation(
                         delay: 1.9,
-                        child: Pinput(
-                          defaultPinTheme: defaultPinTheme,
-                          focusedPinTheme: focusedPinTheme,
-                          submittedPinTheme: submittedPinTheme,
-                          validator: (s) {
-                            return s == '2222' ? null : 'Pin is incorrect';
-                          },
-                          pinputAutovalidateMode:
-                              PinputAutovalidateMode.onSubmit,
-                          showCursor: true,
-                          onCompleted: (pin) {
-                            print(pin);
-                          },
+                        child: const CustomTextFormField(
+                          hinttext: 'New password',
+                          obsecuretext: false,
                         ),
                       ),
                       const SizedBox(
-                        height: 30,
+                        height: 15,
                       ),
                       FadeInAnimation(
                         delay: 2.1,
+                        child: const CustomTextFormField(
+                          hinttext: 'Confirm password',
+                          obsecuretext: false,
+                        ),
+                      ),
+                      SizedBox(
+                        height: 30,
+                      ),
+                      FadeInAnimation(
+                        delay: 2.4,
                         child: CustomElevatedButton(
-                          message: "Verify",
+                          message: "Reset Password ",
                           function: () {
                             GoRouter.of(context)
-                                .pushNamed(Routers.newpassword.name);
+                                .pushNamed(Routers.passwordchanges.name);
                           },
                           color: Colors.black,
                         ),
@@ -122,9 +97,9 @@ class _OtpVerificationPageState extends State<OtpVerificationPage> {
                   ),
                 ),
               ),
-              const Spacer(),
+              Spacer(),
               FadeInAnimation(
-                delay: 2.4,
+                delay: 2.5,
                 child: Padding(
                   padding: const EdgeInsets.only(left: 50),
                   child: Row(
