+++ lib/router/router_config.dart
@@ -1,12 +1,12 @@
-import 'package:authentication_ui/onboarding/onboardingScreen.dart';  // Import the OnboardingScreen
+import 'package:authentication_ui/View/onboarding/onboardingScreen.dart';  // Import the OnboardingScreen
 import 'package:authentication_ui/router/router.dart';
-import 'package:authentication_ui/screens/authentication_ui.dart';
-import 'package:authentication_ui/screens/forget_password.dart';
-import 'package:authentication_ui/screens/login_page.dart';
-import 'package:authentication_ui/screens/new_password.dart';
-import 'package:authentication_ui/screens/otp_verification.dart';
-import 'package:authentication_ui/screens/password_changed.dart';
-import 'package:authentication_ui/screens/signup_page.dart';
+import 'package:authentication_ui/View/screens/authentication_ui.dart';
+import 'package:authentication_ui/View/screens/forget_password.dart';
+import 'package:authentication_ui/View/screens/login_page.dart';
+import 'package:authentication_ui/View/screens/new_password.dart';
+import 'package:authentication_ui/View/screens/otp_verification.dart';
+import 'package:authentication_ui/View/screens/password_changed.dart';
+import 'package:authentication_ui/View/screens/signup_page.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:go_router/go_router.dart';
